Cleanup and color stuff
