+++ .gitignore
@@ -3,4 +3,3 @@
 
 .DS_Store
 yarn.lock
-report.20210711.130815.21384.0.001.json
